Move compat to module_utils/compat/importlib
